remove service worker manager from SW bundle
