+++ src/entries/worker.ts
@@ -1,6 +1,6 @@
 /**
  * New clients will only be including this entry file, which will result in a reduced service worker size.
  */
-import { ServiceWorker } from '../sw/serviceWorker/ServiceWorker';
+import { OneSignalServiceWorker } from '../sw/serviceWorker/ServiceWorker';
 
-(self as any).OneSignal = ServiceWorker;
+(self as any).OneSignal = OneSignalServiceWorker;
