+++ src/shared/context/types.ts
@@ -2,7 +2,8 @@ import type { ISlidedownManager } from 'src/page/managers/slidedownManager/types
 import type { ITagManager } from 'src/page/managers/tagManager/types';
 import type { DynamicResourceLoader } from 'src/page/services/DynamicResourceLoader';
 import type { AppConfig } from '../config/types';
-import type { WorkerMessenger } from '../libraries/WorkerMessenger';
+import type { WorkerMessengerPage } from '../libraries/workerMessenger/page';
+import type { WorkerMessengerSW } from '../libraries/workerMessenger/sw';
 import type PermissionManager from '../managers/PermissionManager';
 import type { ServiceWorkerManager } from '../managers/ServiceWorkerManager';
 import type { ISessionManager } from '../managers/sessionManager/types';
@@ -21,16 +22,15 @@ export interface ContextBase {
 }
 
 export interface ContextSWInterface extends ContextBase {
-  serviceWorkerManager: ServiceWorkerManager<ContextSWInterface>;
   subscriptionManager: SubscriptionManagerSW;
-  workerMessenger: WorkerMessenger<ContextSWInterface>;
+  workerMessenger: WorkerMessengerSW;
   updateManager: UpdateManager<ContextSWInterface>;
 }
 
 export interface ContextInterface extends ContextBase {
-  serviceWorkerManager: ServiceWorkerManager<ContextInterface>;
+  serviceWorkerManager: ServiceWorkerManager;
   subscriptionManager: SubscriptionManagerPage;
-  workerMessenger: WorkerMessenger<ContextInterface>;
+  workerMessenger: WorkerMessengerPage;
   dynamicResourceLoader: DynamicResourceLoader;
   tagManager: ITagManager;
   slidedownManager: ISlidedownManager;
